Gestion des auteurs d'articles

+++ creature_suppression.php
@@ -2,10 +2,23 @@
 include('environnement.php');
 
 $id = htmlspecialchars($_GET['id']);
-
-$request = $bdd->prepare('DELETE FROM creature
+//REQUETE POUR VERIFICATION DU USER A QUI APPARTIENT L'ARTICLE
+$requestSelect = $bdd->prepare('SELECT * FROM creature 
+                                WHERE id=?');
+$requestSelect->execute([$id]);
+while ($data = $requestSelect->fetch()) {
+    //VERIFICATION DU CHAMP USERS_ID AVEC L'ID GARDE EN VARIABLE DE SESSION
+    if ($_SESSION['userId'] == $data['id']) {
+        //ON EXECUTE LA REQUETE SI CA CORRESPOND
+        $request = $bdd->prepare('DELETE FROM creature
                           WHERE id=?');
 
-$request->execute([$id]);
-header('Location: bestiaire.php?success=3');
-exit();
+        $request->execute([$id]);
+        header('Location: bestiaire.php?success=3');
+        exit();
+    } else {
+        //SINON ON RENVOIE SUR L'INDEX
+        header('Location: index.php');
+        exit();
+    }
+}
